# docs/swagger.yaml
swagger: "2.0"
info:
  description: "API for managing companies"
  version: "1.0"
  title: "Company API"
basePath: "/v1"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
...
  /users/:
    get:
      summary: "Search all users"
      tags:
          - users
      operationId: "searchAllUsers"
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"

    post:
      summary: "Create a new user"
      tags:
          - users
      operationId: "createUser"
      parameters:
        - name: "body"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/CreateUserRequest"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/User"
  /users/{id}:
    get:
      summary: "Get a user by ID"
      tags:
          - users
      operationId: "getUser"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "string"
          default: "018cd9aa-546d-793d-8014-728bde77fbed"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/User"
    patch:
      summary: "Update a user by ID"
      tags:
        - users
      operationId: "patchUser"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "string"
          default: "018cd9aa-546d-793d-8014-728bde77fbed"
        - name: "body"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/PatchUserRequest"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/User"
      requestBody:
        content:
          application/json:
            example:
              email: "updateduser@example.com"
              password: "updatedpassword"
              firstName: "Updated"
              lastName: "User"
              mn: "123456789"
              phone: "987654321"
              birthplace: "birthplace"
              companyId: "018cfd19-0d5b-7108-806f-78e1de9fbbcc"
              addressId: "018cfd18-f492-7ade-803a-b414a0a8afac"
              driverId: "018cfd18-f5e6-7fe3-804e-2e9d8152de91"
              dispatcherId: "018cfda7-310b-7f3a-800e-4d9369bd7818"

    delete:
      summary: "Delete a user by ID"
      tags:
        - users
      operationId: "deleteUser"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "string"
          default: "018cf813-4820-7b76-8007-a5f08f9ce129"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/User"
...
  /vehicles/:
    get:
      summary: "Search all vehicles"
      tags:
        - vehicles

      operationId: "searchAllVehicles"
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Vehicle"

    post:
      summary: "Create a new vehicle"
      tags:
        - vehicles
      operationId: "createVehicle"
      parameters:
        - name: "body"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/CreateVehicleRequest"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Vehicle"          
  /vehicles/{id}:
    get:
      summary: "Get a vehicle by ID"
      tags:
        - vehicles
      operationId: "getVehicle"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "string"
          default: 018cd8d8-4f2c-7b70-804e-9fa4d8f22ce8
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Vehicle"
    patch:
      summary: "Update a vehicle by ID"
      tags:
        - vehicles
      operationId: "patchVehicle"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "string"
          default: 018cd8d8-4f2c-7b70-804e-9fa4d8f22ce8
        - name: "body"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/PatchVehicleRequest"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Vehicle"
      requestBody:
        content:
          application/json:
            example:
              plate: "Updated Plate"
              model: "Updated Model"
              producer: "Updated Producer"
              type: "Updated Type"
              weight: 2000
              maxWeight: 3000
              capacity: 5
              homologationMark: "Updated Homologation Mark"
              motorVolume: 2000
              power: 150
              fuel: "Updated Fuel"
              consumption: 10.5
              thrustWeight: 15
              standingSpace: 2
              seatingSpace: 5
              color: "Updated Color"
              shaft: 2
              chasie: "Updated Chasie"
              motor: "Updated Motor"
              insurance: "Updated Insurance"
              registrationDate: "2023-01-01"
              registrationExpiry: "2024-01-01"
              firstRegistrationDate: "2022-01-01"
              ownerMn: 1234567890123
              ownerFirstName: "John"
              ownerLastName: "Doe"
              ownerAddressId: "018cfd18-f492-7ade-803a-b414a0a8afac"
              userFirstName: "Updated User First Name"
              userLastName: "Updated User Last Name"
              userAddressId: "018cfd18-f492-7ade-803a-b414a0a8afac"
              userMn: 9876543210987
              registrationCertificateDate: "2023-01-01"
              registerId: 123
              meta: { "key": "value" }

    delete:
      summary: "Delete a vehicle by ID"
      tags:
        - vehicles
      operationId: "deleteVehicle"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "string"
          default: 018cf7e5-c821-723a-8004-b8cf5b778047
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Vehicle"
  /warrants/:
    get:
      summary: Get all travel warrants
      tags:
        - warrants
      operationId: searchAllWarrant
      produces:
        - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            type: array
            items:
              $ref: "#/definitions/Warrant"

    post:
      summary: Create a travel warrant
      tags:
        - warrants
      operationId: createWarrant
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Warrant object
          required: true
          schema:
            $ref: "#/definitions/CreateWarrantRequest"
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: "#/definitions/Warrant"
  /warrants/{id}:
    get:
      summary: Get a travel warrant by ID
      tags:
        - warrants
      operationId: getWarrant
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of the warrant to get
          required: true
          type: string
          default: "018cd9b6-1486-723b-802e-352321cf6620"
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: "#/definitions/Warrant"

    patch:
      summary: Update a travel warrant by ID
      tags:
        - warrants
      operationId: patchWarrant
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of the warrant to update
          required: true
          type: string
          default: "018cd9b6-1486-723b-802e-352321cf6620"
        - in: body
          name: body
          description: PatchWarrant object
          required: true
          schema:
            $ref: "#/definitions/PatchWarrantRequest"
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: "#/definitions/Warrant"

    delete:
      summary: Delete a travel warrant by ID
      tags:
        - warrants
      operationId: deleteWarrant
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of the warrant to delete
          required: true
          type: string
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: "#/definitions/Warrant"
...
definitions:
  Company:
    type: "object"
    properties:
      id:
        type: "string"
        example: "12345"
      name:
        type: "string"
        example: "Example Company"
      short:
        type: "string"
        example: "ExCo"
      pib:
        type: "string"
        example: "123456789"
      addressId:
        type: "string"
        example: "018cfd18-f492-7ade-803a-b414a0a8afac"
      address:
        $ref: "#/definitions/Address"
      mn:
        type: "string"
        example: "987654321"
      phone:
        type: "string"
        example: "+123456789"
      email:
        type: "string"
        example: "info@example.com"
      employees:
        type: array
        items:
          $ref: "#/definitions/UserCompany"
      meta:
        type: object
        additionalProperties: true
        example: { "key": "value" }
...
  Address:
    type: "object"
    properties:
      id:
        type: "string"
        required: true
        example: "98765"
      city:
        type: "string"
        required: true
        example: "City"
      street:
        type: "string"
        example: "Street"
      streetNo:
        type: "string"
        example: "123"
      name:
        type: "string"
        example: "Company Building"
...
  User:
    type: "object"
    properties:
      id:
        type: "string"
        example: "12345"
      email:
        type: "string"
        example: "user@example.com"
      password:
        type: "string"
        example: "password"
      firstName:
        type: "string"
        example: "John"
      lastName:
        type: "string"
        example: "Doe"
      mn:
        type: "string"
        example: "987654321"
      birthplace:
        type: "string"
        example: "birthplace"
      companies:
        $ref: "#/definitions/UserCompany"
      phone:
        type: "string"
        example: "987654321"  
      addressId:
        type: "string"
        example: "54321"
      address:
        $ref: "#/definitions/Address"
      isDriver:
        type: "bool"
        example: "true"  
      categories:
        $ref: "#/definitions/DriverCategory"
      licenceNumber:
        type: "string"
        example: "12345"
      licenceSerialNumber:
        type: "string"
        example: "54321"
      licenceExpiry:
        type: "string"
        format: "date"
        example: "2023-12-31"
      licenceIssued:
        type: "string"
        format: "date"
        example: "2020-01-01"
      licenceAuthority:
        type: "string"
        example: "Authority"
      roles:
        type: "array"
        items:
          type: "string"
        example: [ "manager", "driver"]  
  PatchUserRequest:
    type: "object"
    properties:
      email:
        type: "string"
        example: "updateduser@example.com"
      password:
        type: "string"
        example: "updatedpassword"
      firstName:
        type: "string"
        example: "Updated"
      lastName:
        type: "string"
        example: "User"
      mn:
        type: "string"
        example: "123456789"
      phone:
        type: "string"
        example: "987654321"  
      addressId:
        type: "string"
        example: "018ce314-6ead-7847-8004-c8e7c8429543"
      isDriver:
        type: "bool"
        example: "false"  
      licenceNumber:
        type: "string"
        example: "123451111"
      licenceSerialNumber:
        type: "string"
        example: "543215555"
      licenceExpiry:
        type: "string"
        format: "date"
        example: "2024-12-31"
      licenceIssued:
        type: "string"
        format: "date"
        example: "2021-01-01"
      licenceAuthority:
        type: "string"
        example: "Updated Authority"
      roles:
        type: array
        items: 
          type: string
        example: ["manager", "driver"]
      addToCompany:
        type: string
        example: 018c6e16-db0d-7f06-800f-fc84904d3f9a
      removeFromCompany:
        type: string
        example: 018c6e16-db0d-7f06-800f-fc84904d3f9a
      status: 
        type: string 
        example: vacation
  CreateUserRequest:
    type: "object"
    required:
      - email
      - password
      - firstName
      - lastName
      - mn
      - companyId
      - addressId

    properties:
      email:
        type: "string"
        example: "newuser@example.com"
      password:
        type: "string"
        example: "newpassword"
      firstName:
        type: "string"
        example: "New"
      lastName:
        type: "string"
        example: "User"
      mn:
        type: "string"
        example: "987654321"
      birthplace:
        type: "string"
        example: "birthplace"
      companyId:
        type: "string"
        example: "018cfd19-0d5b-7108-806f-78e1de9fbbcc"
      phone:
        type: "string"
        example: "987654321"  
      addressId:
        type: "string"
        example: "018cfd18-f492-7ade-803a-b414a0a8afac"
      isDriver:
        type: "bool"
        example: "true"  
      licenceNumber:
        type: "string"
        example: "12345"
      licenceSerialNumber:
        type: "string"
        example: "54321"
      licenceExpiry:
        type: "string"
        format: "date"
        example: "2023-12-31"
      licenceIssued:
        type: "string"
        format: "date"
        example: "2020-01-01"
      licenceAuthority:
        type: "string"
        example: "Authority"
      roles:
        type: "array"
        items:
          type: "string"
        example: [ "manager", "driver"]
  DriverCategory:
    type: "object"
    properties:
      userId:
        type: "string"
        example: sdfsdfdsdfsdf-sdfsdfsdfdsf-sdfdfs-dfs
      category:
        type: string
        example: A
      issued:
        type: string
        format: date
        example: "2023-01-01"
      expired:
        type: string
        format: date
        example: "2024-01-01"
...
  UserCompany:
    type: object
    properties:
      isDeleted:
        type: bool
        example: false
      isActive:
        type: bool
      roles:
        type: "array"
        items:
          type: "string"
        example: [ "manager", "driver"] 
      companyId:
        type: string
        example:  safsdfsadfasd-fadsfsadffds-sdafasd
...
  Vehicle:
    type: object
    properties:
      id:
        type: string
        example: "1"
      plate:
        type: string
        example: "ABC123"
      isTraler:
        type: bool
        example: true
      model:
        type: string
        example: "Sedan"
      producer:
        type: string
        example: "Toyota"
      type:
        type: string
        example: "Car"
      weight:
        type: integer
        example: 1500
      maxWeight:
        type: integer
        example: 2000
      capacity:
        type: integer
        example: 5
      homologationMark:
        type: string
        example: "EU123"
      motorVolume:
        type: integer
        example: 2000
      power:
        type: integer
        example: 150
      fuel:
        type: string
        example: "Gasoline"
      consumption:
        type: number
        example: 8.5
      companyId:
        type: string
        example: asdfsadfdf-asdfsadfsadf-adsfdsfsad
      company:
        $ref: "#/definitions/Company"
      thrustWeight:
        type: integer
        example: 20
      standingSpace:
        type: integer
        example: 2
      seatingSpace:
        type: integer
        example: 3
      color:
        type: string
        example: "Blue"
      shaft:
        type: integer
        example: 2
      chasie:
        type: string
        example: "Chasie123"
      motor:
        type: string
        example: "MotorXYZ"
      insurance:
        type: string
        example: "XYZ123"
      registrationDate:
        type: string
        format: date
        example: "2023-01-01"
      firstRegistrationDate:
        type: string
        format: date-time
        example: "2022-01-01"
      ownerMn:
        type: integer
        example: 123456
      ownerFirstName:
        type: string
        example: John
      ownerLastName:
        type: string
        example: "Doe"
      ownerAddressID:
        type: string
        example: "018cfd18-f492-7ade-803a-b414a0a8afac"
      ownerAddress:
        $ref: "#/definitions/Address"
      userFirstName:
        type: string
        example: "John"
      userLastName:
        type: string
        example: "Doe"
      userAddressID:
        type: string
        example: "018cfd18-f492-7ade-803a-b414a0a8afac"
      userAddress:
        $ref: "#/definitions/Address"
      userMn:
        type: integer
        example: 789012
      registrationCertificateDate:
        type: string
        format: date-time
        example: "2023-01-01"
      registrationExpiry:
        type: string
        format: date-time
        example: "2023-01-01"
      registerId:
        type: integer
        example: 456789
      status:
        type: string
        example: active
      meta:
        type: object
        example: { "key": "value" }
  PatchVehicleRequest:
    type: "object"
    properties:
      plate:
        type: "string"
        example: "Updated Plate"
      model:
        type: "string"
        example: "Updated Model"
      producer:
        type: "string"
        example: "Updated Producer"
      type:
        type: "string"
        example: "Updated Type"
      weight:
        type: "integer"
        example: 2000
      maxWeight:
        type: "integer"
        example: 3000
      capacity:
        type: "integer"
        example: 5
      homologationMark:
        type: "string"
        example: "Updated Homologation Mark"
      motorVolume:
        type: "integer"
        example: 2000
      power:
        type: "integer"
        example: 150
      fuel:
        type: "string"
        example: "Updated Fuel"
      consumption:
        type: "number"
        example: 10.5
      thrustWeight:
        type: "integer"
        example: 15
      standingSpace:
        type: "integer"
        example: 2
      seatingSpace:
        type: "integer"
        example: 5
      color:
        type: "string"
        example: "Updated Color"
      shaft:
        type: "integer"
        example: 2
      chasie:
        type: "string"
        example: "Updated Chasie"
      motor:
        type: "string"
        example: "Updated Motor"
      insurance:
        type: "string"
        example: "Updated Insurance"
      registrationDate:
        type: "string"
        format: "date"
        example: "2023-01-01"
      firstRegistrationDate:
        type: "string"
        format: "date"
        example: "2022-01-01"
      ownerMn:
        type: "integer"
        example: 1234567890123
      ownerFirstName:
        type: "string"
        example: "John"
      ownerLastName:
        type: "string"
        example: "Doe"
      ownerAddressId:
        type: "string"
        example: "018cfd18-f492-7ade-803a-b414a0a8afac"
      userFirstName:
        type: "string"
        example: "Updated User First Name"
      userLastName:
        type: "string"
        example: "Updated User Last Name"
      userAddressId:
        type: "string"
        example: "018cfd18-f492-7ade-803a-b414a0a8afac"
      userMn:
        type: "integer"
        example: 9876543210987
      registrationCertificateDate:
        type: "string"
        format: "date"
        example: "2023-01-01"
      registrationExpiry:
        type: "string"
        format: "date"
        example: "2023-01-01"
      registerId:
        type: "integer"
        example: 123 
      meta:
        type: "object"
        additionalProperties:
          type: "string"
  CreateVehicleRequest:
    type: "object"
    properties:
      plate:
        type: "string"
        example: "New Plate"
      model:
        type: "string"
        example: "New Model"
      producer:
        type: "string"
        example: "New Producer"
      type:
        type: "string"
        example: "New Type"
      weight:
        type: "integer"
        example: 2000
      maxWeight:
        type: "integer"
        example: 3000
      capacity:
        type: "integer"
        example: 5
      homologationMark:
        type: "string"
        example: "New Homologation Mark"
      motorVolume:
        type: "integer"
        example: 2000
      power:
        type: "integer"
        example: 150
      fuel:
        type: "string"
        example: "New Fuel"
      consumption:
        type: "number"
        example: 10.5
      thrustWeight:
        type: "integer"
        example: 15
      standingSpace:
        type: "integer"
        example: 2
      seatingSpace:
        type: "integer"
        example: 5
      color:
        type: "string"
        example: "New Color"
      shaft:
        type: "integer"
        example: 2
      chasie:
        type: "string"
        example: "New Chasie"
      motor:
        type: "string"
        example: "New Motor"
      insurance:
        type: "string"
        example: "New Insurance"
      registrationDate:
        type: "string"
        format: "date"
        example: "2023-01-01"
      registrationExpiry:
        type: "string"
        format: "date"
        example: "2024-01-01"  
      firstRegistrationDate:
        type: "string"
        format: "date"
        example: "2022-01-01"
      ownerMn:
        type: "integer"
        example: 1234567890123
      ownerLastName:
        type: "string"
        example: "New Owner Last Name"
      ownerAddressId:
        type: "string"
        example: "018cfd18-f492-7ade-803a-b414a0a8afac"
      userFirstName:
        type: "string"
        example: "New User First Name"
      userLastName:
        type: "string"
        example: "New User Last Name"
      userAddressId:
        type: "string"
        example: "018cfd18-f492-7ade-803a-b414a0a8afac"
      userMn:
        type: "integer"
        example: 9876543210987
      registrationCertificateDate:
        type: "string"
        format: "date"
        example: "2023-01-01"
      registerId:
        type: "integer"
        example: 123
      meta:
        type: "object"
        additionalProperties:
          type: "string"
  WarrantStatus:
    type: string
    enum:
      - preparation
      - assigned
      - transit
      - transport
      - interrupted
      - completed
  Warrant:
    type: object
    properties:
      id:
        type: string
        example: "1"
      issueDate:
        type: string
        format: date-time
        example: "2023-12-31 12:00:00"
      expectedStart:
        type: string
        format: date-time
        example: "2023-12-30 12:00:00"
      closingDate:
        type: string
        format: date-time
        example: "2023-12-31 12:00:00"
      driverId:
        type: string
        example: "018cfd18-f5e6-7fe3-804e-2e9d8152de91"
      driver:
        $ref: "#/definitions/User"
      passengers:
        type: array
        items: 
          type: string
        example: ["Vasily Zaitzev", "Luke Skywalker"]
      vehicleId:
        type: string
        example: "018cfd19-0da3-727a-806c-9293bb95c4cf"
      vehicle:
        $ref: "#/definitions/Vehicle"
      trailerId:
        type: string
        example: "018cfd19-0da3-727a-806c-9293bb95c4cf"
      trailer:
        $ref: "#/definitions/Vehicle"
      companyId:
        type: string
        example: "018cfd19-0d5b-7108-806f-78e1de9fbbcc"
      company:
        $ref: "#/definitions/Company"
      dispatcherId:
        type: string
        example: "018cfda7-310b-7f3a-800e-4d9369bd7818"
      dispatcher:
        $ref: "#/definitions/User"
      status:
        $ref: "#/definitions/WarrantStatus"
      name: 
        type: string
        example: redovna emezeta
      note:
        type: string
        example: vozac je imao zdravtvenih problema zadnje vreme
  CreateWarrantRequest:
    type: object
    properties:
      expectedStart:
        type: string
        format: date-time
        example: "2023-01-01 15:04:05"
      driverId:
        type: string
        example: "018cfd18-f5e6-7fe3-804e-2e9d8152de91"
      vehicleId:
        type: string
        example: "018cfd19-0da3-727a-806c-9293bb95c4cf"
      companyId:
        type: string
        example: "018cfd19-0d5b-7108-806f-78e1de9fbbcc"
      dispatcherId:
        type: string
        example: "018cfda7-310b-7f3a-800e-4d9369bd7818"
      technicalCorrectnes:
        type: string
        example: ready
      passangers: 
        type: array
        items:
          type: string
        example: 
          example: [ "Peter Zaitzev", "Pit Murich"]  
      status:
        type: string
        example: "assigned"  
      name:
        type: string  
        example: redovna emezata
      note: 
        type: string
        example: Proveri da li se vozilo moze prikljuciti na ulazana vrata magacina
  PatchWarrantRequest:
    type: object
    properties:
      expectedStart:
        type: string
        format: date-time
        example: "2023-01-01 15:04:05"
      driverId:
        type: string
        example: "018cfd18-f5e6-7fe3-804e-2e9d8152de91"
      vehicleId:
        type: string
        example: "018cfd19-0da3-727a-806c-9293bb95c4cf"
      companyId:
        type: string
        example: "018cfd19-0d5b-7108-806f-78e1de9fbbcc"
      dispatcherId:
        type: string
        example: "018cfda7-310b-7f3a-800e-4d9369bd7818"
      technicalCorrectnes:
        type: string
        example: ready
      name:
        type: string  
        example: redovna emezata
      note: 
        type: string
        example: Proveri da li se vozilo moze prikljuciti na ulazana vrata magacina
      addPassenger:
        type: string  
        example: Pit Murich
      removePassenger: 
        type: string
        example: Don Jonson
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer